Split test_autocast_modules.py into separate test files to mirror the source file structure.
